# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Villa Access Control Platform API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Villa:
      type: object
      required: [id, name, gate_id, door_id, status]
      properties:
        id:
          type: string
        name:
          type: string
        gate_id:
          type: string
        door_id:
          type: string
        camera_ids:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, maintenance]
    VillaInput:
      type: object
      required: [name, gate_id, door_id]
      properties:
        name:
          type: string
        gate_id:
          type: string
        door_id:
          type: string
        camera_ids:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, maintenance]
...
paths:
...
  /villas:
    get:
      operationId: getVillas
      tags: [villas]
      summary: List all villas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of villas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Villa'
    post:
      operationId: createVilla
      tags: [villas]
      summary: Create a new villa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VillaInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Villa'
  /villas/{id}:
    get:
      operationId: getVilla
      tags: [villas]
      summary: Get a villa by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Villa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Villa'
    put:
      operationId: updateVilla
      tags: [villas]
      summary: Update a villa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VillaInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Villa'
...